% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_parameter.R
\name{analysis.amount}
\alias{analysis.amount}
\title{Estimate Analysis Fluid consumption}
\usage{
analysis.amount(sample.duration, medium.flow, unit.flow = "mL/min")
}
\arguments{
\item{sample.duration}{Numeric value representing the duration of each sample in minutes.}

\item{samples.to.go}{Integer representing the number of remaining samples.}

\item{time}{The starting time of analysis, default is the current system time.}

\item{unit}{Time unit for the output, can be "h" for hours or "min" for minutes. Default is "h".}
}
\value{
A message string describing the expected completion time.
}
\description{
This function calculates the estimated medium amount for ongoing sample analysis, given the flow of a medium and the duration of the total analysis.
}
