% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_analysis_time_amount.R
\name{analysis_time}
\alias{analysis_time}
\title{Estimate Analysis Completion Time}
\usage{
analysis_time(
  sample.duration,
  samples.to.go,
  time = Sys.time(),
  unit = c("h", "min")
)
}
\arguments{
\item{sample.duration}{Numeric value representing the duration of each sample in minutes.}

\item{samples.to.go}{Integer representing the number of remaining samples.}

\item{time}{The starting time of analysis, default is the current system time.}

\item{unit}{Time unit for the output, can be "h" for hours or "min" for minutes. Default is "h".}
}
\value{
A message string describing the expected completion time.
}
\description{
This function calculates the estimated completion time for ongoing sample analysis, given the duration of each sample and the number of samples remaining.
}
