% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_calc.R
\name{calibration.calc}
\alias{calibration.calc}
\title{Calibration Calculation}
\usage{
calibration.calc(
  area,
  calibration_model = cal.file,
  substance,
  dilution = 1,
  sub = NA
)
}
\arguments{
\item{area}{Numeric value, the observed area or height from the instrument.}

\item{calibration_model}{calibration model from rds file.}

\item{substance}{Character, the name of the substance being analyzed.}

\item{dilution}{Numeric, the dilution factor applied to the sample. Default is 1.}

\item{sub}{Optional, a matrix specifying which subset of calibration data to use for recalculating the model.}
}
\value{
Numeric value, the calculated final concentration of the substance.
}
\description{
This function calculates the final concentration of a substance based on calibration models
from instruments like IC, HPLC, or GC-MS.
}
\examples{
\dontrun{
final_conc <- calibration.calc(area = 1500, instrument = "IC", substance = "NaCl", date = "2024-01-01")
}
}
