% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_file.R
\name{calibration.matrix}
\alias{calibration.matrix}
\title{Generate a Calibration Matrix from a Calibration File}
\usage{
calibration.matrix(cal.file = cal.file)
}
\arguments{
\item{cal.file}{A list of calibration files (models) loaded from \code{.rds} files. Each element of the list should contain a model component.}
}
\value{
A matrix where each row corresponds to a data point and each column represents a different calibration model.
}
\description{
This function generates a calibration matrix from a list of calibration models. The matrix is constructed from the model data of the calibration files.
}
\examples{
\dontrun{
# Generate a calibration matrix from a list of calibration models
calibration_data <- calibration.file(instrument = "IC", type = "area", date = "2024-07-01")
cal_matrix <- calibration.matrix(cal.file = calibration_data)
}

}
