% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_factor.R
\name{chemical_factor}
\alias{chemical_factor}
\title{Create a Factor for Chemicals Based on Retention Time and a Reference Sheet}
\usage{
chemical_factor(chemicals, retention.time = NA, xlsx = chemicals.master)
}
\arguments{
\item{chemicals}{A character vector representing the list of chemical compounds.}

\item{retention.time}{Numeric, optional. Retention times for the chemicals. Default is NA.}

\item{xlsx}{A data frame or list (Excel sheet) containing chemical properties such as Formula, Compound, German Name, Compound_Alternative, and color. Default is \code{chemicals.master}.}
}
\value{
A factor where the levels are ordered based on the reference sheet or retention time.
}
\description{
This function creates a factor for a list of chemicals, optionally ordering them based on retention time. It matches the chemicals with a reference sheet (Excel) that contains various chemical properties.
If a chemical is not found in the reference, a warning is issued and the chemical is placed at the end.
}
\examples{
\dontrun{
chemicals <- c("H2O", "NaCl", "CO2")
retention.time <- c(2.5, 1.3, 3.1)
chemical_factor(chemicals, retention.time, xlsx = chemicals.master)
}

}
