% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_mol_einwaage.R
\name{convert_mol_to_Einwaage}
\alias{convert_mol_to_Einwaage}
\title{Convert Molar Concentration to Weight (Einwaage)}
\usage{
convert_mol_to_Einwaage(
  SOLL_concentration = c(1, 50),
  SOLL_unit = "mmol/L",
  return_unit = "g",
  Volume = 2,
  Volume_unit = "L",
  substance = "Na2CO3",
  molar_mass = r4apl$chem$molar_mass[[which(names(r4apl$chem$molar_mass) \%in\%
    substance)]]
)
}
\arguments{
\item{SOLL_concentration}{Numeric vector, the target concentration in mol/L or mmol/L.}

\item{SOLL_unit}{Character, the unit of the target concentration. Default is "mmol/L".}

\item{return_unit}{Character, the desired unit for the resulting weight. Options are "g" or "mg". Default is "g".}

\item{Volume}{Numeric, the volume of the solution. Default is 2.}

\item{Volume_unit}{Character, the unit of the volume. Options are "L" or "mL". Default is "L".}

\item{substance}{Character, the name of the substance for which the weight is being calculated. Default is "Na2CO3".}

\item{molar_mass}{Numeric, the molar mass of the substance. Default is obtained from the r4apl package.}
}
\value{
A data frame containing the concentration, unit, and calculated weight (Einwaage).
}
\description{
This function converts a specified molar concentration (e.g., mmol/L or mol/L) to a weight (Einwaage) based on the molar mass of a substance, volume, and the desired return unit (e.g., grams or milligrams).
}
\examples{
\dontrun{
convert_mol_to_Einwaage(SOLL_concentration = 10, SOLL_unit = "mmol/L", Volume = 1, substance = "NaCl")
}

}
