% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{datetime}
\alias{datetime}
\title{Format Datetime String}
\usage{
datetime(datetime = Sys.time(), time = FALSE)
}
\arguments{
\item{datetime}{A datetime object. Default is the current system time.}

\item{time}{Logical, if TRUE, includes both date and time in the output. Default is FALSE (date only).}
}
\value{
A formatted string representing the date or datetime.
}
\description{
This function formats a datetime object, either extracting the date portion or both date and time, with customizable formatting.
}
