% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Load Initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library name.}

\item{pkgname}{The package name.}
}
\value{
TRUE (ensures the function completes successfully).
}
\description{
This function runs automatically when the package is loaded, displaying version information and additional details.
}
