% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_propagation.R
\name{error_propagation_division}
\alias{error_propagation_division}
\title{Error Propagation for Division of Two Quantities}
\usage{
error_propagation_division(mean_A, sd_A, mean_B, sd_B)
}
\arguments{
\item{mean_A}{Numeric, the mean value of quantity A.}

\item{sd_A}{Numeric, the standard deviation (error) of quantity A.}

\item{mean_B}{Numeric, the mean value of quantity B.}

\item{sd_B}{Numeric, the standard deviation (error) of quantity B.}
}
\value{
A list containing the ratio of the means (C = mean_A / mean_B) and the propagated error (sigma_C).
}
\description{
This function calculates the propagated error when dividing two quantities, taking into account their means and standard deviations.
}
\examples{
\dontrun{
result <- error_propagation_division(mean_A = 10, sd_A = 0.5, mean_B = 5, sd_B = 0.2)
print(result)
}

}
