% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{extract_and_rescale_pca_arrows}
\alias{extract_and_rescale_pca_arrows}
\title{Extract and Rescale PCA Arrows for Visualization}
\usage{
extract_and_rescale_pca_arrows(pca_model, pc_x = 1, pc_y = 2, scale_factor = 5)
}
\arguments{
\item{pca_model}{A PCA model object containing the loadings and explained variance.}

\item{pc_x}{Integer, the first principal component for the x-axis. Default is 1.}

\item{pc_y}{Integer, the second principal component for the y-axis. Default is 2.}

\item{scale_factor}{Numeric, the factor by which to rescale the loadings. Default is 5.}
}
\value{
A data frame containing the variable names and the rescaled loadings for the specified principal components.
}
\description{
This function extracts the loadings from a PCA model and rescales them for better visualization of PCA arrows. It calculates the correlation loadings for the first two principal components and rescales them by a specified factor.
}
\examples{
\dontrun{
extract_and_rescale_pca_arrows(pca_model = powerco$pca$GC, pc_x = 1, pc_y = 2, scale_factor = 5)
}

}
