% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_read_GC_chromtogram.R
\name{find.read.GC.chromatogram}
\alias{find.read.GC.chromatogram}
\title{Read GC Chromatogram and Detector Data}
\usage{
find.read.GC.chromatogram(database = GC_MS$database)
}
\arguments{
\item{database}{A data frame or list containing details about the GC data, including the directory paths (\code{dir}), file basenames (\code{basename}), and detector types (\code{detector}).}
}
\value{
A list where each element corresponds to a detector type, containing the raw data from the chromatogram files.
The data is organized by detector and filename.
}
\description{
This function reads Gas Chromatography (GC) chromatogram and Druck data from CSV files, based on the entries in the provided database.
It filters files by the specified sample ID and detector types and returns the raw data for each detector.
}
\examples{
\dontrun{
# Assuming GC_MS$database is a data frame with the necessary fields
gc_data <- find.read.GC.chromatogram(database = GC_MS$database)
}

}
