% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_read_GC_database.R
\name{find.read.GC.database}
\alias{find.read.GC.database}
\title{Read GC Database and Filter Data}
\usage{
find.read.GC.database(
  dir = c(wd$GC_MS$data, wd$GC_FID_TCD$data),
  date = NA,
  time = NA,
  order = NA,
  subdate = "2024-07-01",
  instrument = "GC_MS"
)
}
\arguments{
\item{dir}{Character vector, the directories where GC database files are stored. Default is \code{c(wd$GC_MS$data, wd$GC_FID_TCD$data)}.}

\item{date}{Character, optional. A manual date (format: "YYYYMMDD") to filter the files. Default is NA.}

\item{time}{Character, optional. A manual time (format: "HHMM") to further filter the files by time. Default is NA.}

\item{order}{Integer, optional. Specifies the order position of files from newest to oldest. Default is NA.}

\item{subdate}{Character, optional. A date (format: "YYYY-MM-DD") to filter the data by the \code{datetime} field. Default is "2024-07-01".}

\item{instrument}{Character, specifies the instrument type to filter directories, either "GC_MS" or "GC_FID_TCD". Default is "GC_MS".}
}
\value{
A \code{data.table} containing the filtered GC data.
}
\description{
This function reads Gas Chromatography (GC) database files from a specified directory, allows filtering by date and time, and optionally returns a subset of the data.
}
\examples{
\dontrun{
# Read GC database and filter by date and time
find.read.GC.database(date = "20240701", time = "1200")
}

}
