% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_read_IC_chromtogram.R
\name{find.read.IC.chromatogram}
\alias{find.read.IC.chromatogram}
\title{Read IC Chromatogram and Druck Data}
\usage{
find.read.IC.chromatogram(ID)
}
\arguments{
\item{ID}{Character, the sample ID to search for in the text files.}
}
\value{
A list containing the datetime, identification (Ident), sample ID, and optionally the Druck and Chromatogram data if found.
}
\description{
This function reads Ion Chromatography (IC) chromatogram and Druck data from a text file, based on the specified sample ID.
}
\examples{
\dontrun{
find.read.IC.chromatogram(ID = "20240701_1234")
}

}
