% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_read_IC_database.R
\name{find.read.IC.database}
\alias{find.read.IC.database}
\title{Read IC Database and Filter Data}
\usage{
find.read.IC.database(
  dir = wd$team_labor$datensicherung$IC,
  date = NA,
  order = NA,
  subdat = TRUE,
  subdate = "2024-07-01"
)
}
\arguments{
\item{dir}{Character, the directory where IC database files are stored. Default is \code{wd$team_labor$datensicherung$IC}.}

\item{date}{Character, optional. A manual date (format: "YYYYMMDD") to filter the files. Default is NA.}

\item{order}{Integer, optional. Specifies the order position of files from newest to oldest. Default is NA.}

\item{subdat}{Logical, whether to return a smaller subset of the data. Default is TRUE.}

\item{subdate}{Character, a date (format: "YYYY-MM-DD") to filter the data by \code{Bestimmungsstart}. Default is "2024-07-01".}

\item{time}{Character, optional. A manual time (format: "HHMM") to filter the files. Default is NA.}
}
\value{
A data.table containing the filtered IC data.
}
\description{
This function reads Ion Chromatography (IC) database files from a specified directory, allows filtering by date and time, and optionally returns a subset of the data.
}
\examples{
\dontrun{
find.read.IC.database(date = "20240701", time = "1200")
}

}
