% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Sysgetenv_folder.R
\name{find.Sys.getenv.folder}
\alias{find.Sys.getenv.folder}
\title{Find Folder in System Environment Path}
\usage{
find.Sys.getenv.folder(
  x = "LOCALAPPDATA",
  pattern = "Teckso",
  pattern2 = "GmbH"
)
}
\arguments{
\item{x}{Character, the name of the system environment variable to search within (e.g., "LOCALAPPDATA"). Default is "LOCALAPPDATA".}

\item{pattern}{Character, the first pattern to match folder names against. Default is "Teckso".}

\item{pattern2}{Character, an optional second pattern to narrow down results if multiple folders match the first pattern. Default is "GmbH".}
}
\value{
A string representing the path to the matched folder. If no folder is found, a warning is issued.
}
\description{
This function searches for a folder within a system environment path (e.g., "LOCALAPPDATA") based on specified patterns.
It can handle multiple matches and returns the first matching folder.
}
\examples{
\dontrun{
find.Sys.getenv.folder(x = "LOCALAPPDATA", pattern = "Teckso", pattern2 = "GmbH")
}

}
