% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{group_peaks_global}
\alias{group_peaks_global}
\title{Group Peaks Globally by RT Clustering}
\usage{
group_peaks_global(
  integration.results,
  rt_min = 3,
  rt_max = 11,
  h = 0.25,
  prefix = "G1"
)
}
\arguments{
\item{integration.results}{Named list of data.frames with integrated peak info}

\item{rt_min}{Lower RT limit}

\item{rt_max}{Upper RT limit}

\item{h}{Height parameter for clustering (see cutree)}

\item{prefix}{Prefix for group name}
}
\value{
Data.frame with compound_group and median_RT
}
\description{
Clusters peaks globally based on RT similarity using hierarchical clustering
}
