% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack.R
\name{hack}
\alias{hack}
\title{Calculate Working Hours and Remaining Time}
\usage{
hack(start_time, ist_time = Sys.time(), pause_time = NA, Zeitkonto = NA)
}
\arguments{
\item{start_time}{Character, the starting time of the workday (format: "HH:MM").}

\item{ist_time}{Character or POSIXct, the current or end time of the workday. Default is \code{Sys.time()}.}

\item{pause_time}{Numeric, the length of the break (pause) in minutes. Default is NA.}

\item{Zeitkonto}{Numeric, the current "Zeitkonto" (time account balance). Default is NA.}
}
\value{
A data frame with the calculated times and remaining work time.
}
\description{
This function calculates the total working hours, remaining time until a full workday (8 or 9 hours), and updates the "Zeitkonto" (time account) if provided.
}
\examples{
\dontrun{
hack(start_time = "08:00", ist_time = "16:30", pause_time = 30, Zeitkonto = 1.5)
}

}
