% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{head10}
\alias{head10}
\title{Display a Subset of Data}
\usage{
head10(x, nrowp = 10, ncolp = 1:10, tail = FALSE)
}
\arguments{
\item{x}{A data frame or data table.}

\item{nrowp}{Number of rows to display. Default is 10.}

\item{ncolp}{A vector of column indices to display, default is 1:10.}

\item{tail}{Logical, if TRUE, shows the last rows, otherwise shows the first rows. Default is FALSE.}
}
\value{
A subset of the data frame or data table.
}
\description{
Displays the first or last 10 rows and selected columns from a data frame or data table.
}
