% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_peaks}
\alias{integrate_peaks}
\title{Trapezoidal Peak Integration}
\usage{
integrate_peaks(x, y)
}
\arguments{
\item{x}{Numeric vector for x-axis (e.g. RT)}

\item{y}{Numeric vector for y-axis (e.g. intensity)}
}
\value{
Numeric value representing the area under the curve
}
\description{
Integrates a signal using the trapezoidal rule
}
