% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interflex.R
\name{interflex.Nebenzeiteneintrag}
\alias{interflex.Nebenzeiteneintrag}
\title{Process Nebenzeiteneintrag Entries in Interflex Files}
\usage{
interflex.Nebenzeiteneintrag(
  interflex.files = interflex.files,
  month = month(Sys.Date()),
  year = year(Sys.Date())
)
}
\arguments{
\item{interflex.files}{A vector of Interflex file names.}

\item{month}{The month for which to extract Nebenzeiteneintrag. Default is the current month.}

\item{year}{The year for which to extract Nebenzeiteneintrag. Default is the current year.}
}
\value{
A list containing processed Nebenzeiteneintrag data, including the date and day of the week.
}
\description{
This function processes Nebenzeiteneintrag (additional time) entries from the Interflex files for a specific month and year.
}
