% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_interflex.R
\name{interflex_Clipboard}
\alias{interflex_Clipboard}
\title{Process Interflex Data from Clipboard}
\usage{
interflex_Clipboard(dat.cb = readClipboard())
}
\arguments{
\item{dat.cb}{A character vector containing data from the clipboard. Default is to read from the clipboard.}
}
\value{
A data frame with processed Interflex data.
}
\description{
This function processes data copied from the clipboard (Interflex) and extracts relevant details such as dates, working times, and others.
}
