% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack_interflex.R
\name{interflex_Zeitkto_Beginn}
\alias{interflex_Zeitkto_Beginn}
\title{Rename Interflex Zeitkonto Files}
\usage{
interflex_Zeitkto_Beginn(dat1 = interflex.out)
}
\arguments{
\item{dat1}{A data object that contains the 'Zeitkto' (time account) and 'Beginn' (start time) fields used for renaming.}
}
\value{
This function renames files in the specified directory but does not return a value.
}
\description{
This function renames Interflex Zeitkonto (time account) files to a new naming convention based on provided data.
The function changes the name of files from "IST" (current state) to "SOLL" (target state) with the correct time.
}
\details{
The function looks for files in the directory specified by \code{wd$R$master}, and renames files following the pattern
"interflex_Zeitkto_IST" to the format "interflex_Zeitkto_\if{html}{\out{<Zeitkto>}}\emph{Beginn}\if{html}{\out{<Time>}}" where \code{Zeitkto} and \code{Beginn} come
from the provided \code{dat1} argument.
}
\examples{
\dontrun{
interflex_Zeitkto_Beginn(dat1 = interflex.out)
}

}
