% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{kw}
\alias{kw}
\title{Get ISO Week Number}
\usage{
kw(date = Sys.Date())
}
\arguments{
\item{date}{A date object. Default is the current system date.}
}
\value{
An integer representing the ISO week number.
}
\description{
This function returns the ISO week number for a given date.
}
