% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{moveme}
\alias{moveme}
\title{Reorder Elements in a Vector}
\usage{
moveme(invec, movecommand)
}
\arguments{
\item{invec}{A character vector representing the original order of elements.}

\item{movecommand}{A string specifying how elements should be moved, using keywords such as "before", "after", "first", and "last".}
}
\value{
A reordered character vector based on the move command.
}
\description{
This function moves elements within a vector based on a specified move command.
}
