% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_mz_identify.R
\name{mz_identify}
\alias{mz_identify}
\title{Identify Substances Based on m/z Value}
\usage{
mz_identify(mz)
}
\arguments{
\item{mz}{A numeric value representing the m/z (mass-to-charge ratio) to identify.}
}
\value{
A string or an expression representing the identified substance or chemical species based on the m/z value.
If the m/z is not recognized, it returns "unknown substance".
}
\description{
This function returns an expression representing a chemical formula or a substance name based on a given mass-to-charge ratio (m/z).
}
\examples{
mz_identify(1)   # Returns "H"^"+"
mz_identify(32)  # Returns "O"[2]
mz_identify(99)  # Returns "unknown substance"

}
