% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{opendir}
\alias{opendir}
\title{Open a Directory in the File Explorer or Browser}
\usage{
opendir(dir = getwd())
}
\arguments{
\item{dir}{A string specifying the directory to open. Default is the current working directory.}
}
\description{
This function opens a specified directory in the system's file explorer (Windows) or in the default web browser (non-Windows systems).
}
