% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{parmfrow}
\alias{parmfrow}
\title{Determine Rows and Columns for Plot Layout}
\usage{
parmfrow(n)
}
\arguments{
\item{n}{Integer, the number of plots.}
}
\value{
A vector of two integers: the number of rows and columns for the plot layout.
}
\description{
This function calculates the optimal number of rows and columns for a plotting layout based on the total number of plots.
}
