% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.lab}
\alias{pca.lab}
\title{Generate PCA Component Label with Explained Variance}
\usage{
pca.lab(pca = powerco$pca$GC, ncomp = 1)
}
\arguments{
\item{pca}{A PCA model object containing the results of the PCA analysis.}

\item{ncomp}{Integer, the component number to label. Default is 1.}
}
\value{
A character string with the component label and explained variance in percentage.
}
\description{
This function generates a label for a principal component (PC) that includes the explained variance for a specified component.
}
\examples{
\dontrun{
pca.lab(pca = powerco$pca$GC, ncomp = 1)
}

}
