% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_IC_chromatogram.R
\name{process_section_ic}
\alias{process_section_ic}
\title{Helper Function to Process Sections in the IC File}
\usage{
process_section_ic(raw_data, section_name, end_marker = NULL)
}
\arguments{
\item{raw_data}{The raw data read from the txt file.}

\item{section_name}{The section name to process (e.g., "Anionen", "Druck").}

\item{end_marker}{The marker indicating the end of the section. Default is NULL (process until the end).}
}
\value{
A data.table containing the processed section.
}
\description{
This helper function processes sections of the IC file (e.g., Chromatogram or Druck) based on the given section name.
}
