% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{rangexy}
\alias{rangexy}
\title{Adjust Range of Data with Percentile Changes}
\usage{
rangexy(xy, p = c(5, 5))
}
\arguments{
\item{xy}{A numeric vector or matrix of data values.}

\item{p}{A two-element numeric vector specifying the percentage adjustments for the lower and upper limits. Default is c(5, 5).}
}
\value{
A numeric vector of length 2 representing the adjusted lower and upper limits of the range.
}
\description{
This function calculates the range of a numeric vector and adjusts its limits based on percentage changes specified by the user.
}
