% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teckso_gas_sampler_terminal.R
\name{read.teckso.log}
\alias{read.teckso.log}
\title{Read and Process Teckso Log File}
\usage{
read.teckso.log(logfile, header = FALSE, sep = "\\n", col.names = "log_entry")
}
\arguments{
\item{logfile}{A string representing the path to the log file.}

\item{header}{Logical, indicating whether the file contains a header. Default is FALSE.}

\item{sep}{A string representing the separator used in the log file. Default is newline (\code{"\\n"}).}

\item{col.names}{A string or vector of strings representing column names for the log entries. Default is "log_entry".}
}
\value{
A list containing two data frames: one for pressure readings of P1 and one for pressure readings of P2.
Each data frame contains datetime, pressure, and unit columns.
}
\description{
This function reads a Teckso log file, extracts relevant information such as timestamps, elapsed time, and messages,
and specifically processes pressure logs for P1 and P2 (in mBar). It returns pressure readings along with their corresponding timestamps.
}
\examples{
\dontrun{
log_data <- read.teckso.log(logfile = "path/to/teckso.log")
}

}
