% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_IC_chromatogram.R
\name{read_ic_chromatogram}
\alias{read_ic_chromatogram}
\title{Read IC Chromatogram and Druck Data}
\usage{
read_ic_chromatogram(ID)
}
\arguments{
\item{ID}{Character, the sample ID to search for in the text files.}
}
\value{
A list containing the datetime, identification (Ident), sample ID, and optionally the Druck and Chromatogram data if found.
}
\description{
This function reads Ion Chromatography (IC) chromatogram and Druck data from a text file, based on the specified sample ID.
}
\examples{
\dontrun{
read_ic_chromatogram(ID = "20240701_1234")
}

}
