% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rect.R
\name{rectplot}
\alias{rectplot}
\title{Create a Rectangular Plot with Segments}
\usage{
rectplot(
  y1,
  y2,
  angle = 45,
  lty = 1,
  lwd = 1,
  density = 0,
  col = NA,
  width = 0.25,
  border = "black"
)
}
\arguments{
\item{y1}{Numeric vector, the heights of the bars.}

\item{y2}{Numeric vector, the additional heights to be added as segments on top of the bars.}

\item{angle}{Numeric, the angle of shading lines. Default is 45.}

\item{lty}{Integer, line type for the borders of the rectangles. Default is 1 (solid).}

\item{lwd}{Numeric, line width for the borders of the rectangles and segments. Default is 1.}

\item{density}{Numeric, the density of shading lines, if shading is applied. Default is 0 (no shading).}

\item{col}{Color, the fill color of the rectangles. Default is NA (no fill).}

\item{width}{Numeric, the width of the bars. Default is 0.25.}

\item{border}{Color, the color of the rectangle borders and segments. Default is "black".}
}
\value{
A sequence corresponding to the x positions of the bars, for further customization if needed.
}
\description{
This function generates a bar plot with rectangular bars, overlaying segments for additional data visualization. It is useful for plotting two sets of related data.
}
\examples{
\dontrun{
y1 <- c(3, 5, 2, 6)
y2 <- c(1, 2, 3, 1)
rectplot(y1, y2, col = "blue", border = "red", lwd = 2)
}

}
