% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interflex.R
\name{replace_german_days}
\alias{replace_german_days}
\title{Replace German Day Abbreviations with English Equivalents}
\usage{
replace_german_days(date_string)
}
\arguments{
\item{date_string}{A character string that contains German day abbreviations.}
}
\value{
A character string with German day abbreviations replaced by their English equivalents.
}
\description{
This function replaces German day abbreviations (e.g., "Mo.") with their English equivalents (e.g., "Mon").
}
\examples{
replace_german_days("Mo. 12")  # returns "Mon 12"

}
