% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{subscript_formula}
\alias{subscript_formula}
\title{Create a Subscripted Formula}
\usage{
subscript_formula(formula, bold = TRUE)
}
\arguments{
\item{formula}{A string representing the formula.}

\item{bold}{Logical, if TRUE, makes the base part bold. Default is TRUE.}
}
\value{
A parsed expression suitable for rendering in plots or other mathematical displays.
}
\description{
This function takes a formula and adds subscripts to any numeric values found at the end.
Optionally, the base of the formula can be bolded.
}
