% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_gc.R
\name{transfer.ms}
\alias{transfer.ms}
\title{Transfer Mass Spectrometry Data}
\usage{
transfer.ms(dat, mass.pattern = "m[0-9]", time.pattern = "time_MS")
}
\arguments{
\item{dat}{A data frame or data table containing MS data with columns for mass and time.}

\item{mass.pattern}{A regular expression pattern to identify mass-related columns in \code{dat}. Default is \code{"m[0-9]"}.}

\item{time.pattern}{A string to identify the time-related column in \code{dat}. Default is \code{"time_MS"}.}
}
\value{
A list containing:
\item{time}{Time values extracted from the data.}
\item{mass}{Mass values extracted and processed from the data.}
\item{mz}{The mass/charge (m/z) ratios from the mass-related columns.}
\item{mz.mtx}{A matrix of mass/charge values.}
\item{mz.log10.p1.mtx}{A matrix of log10-transformed m/z values (log10(m/z + 1)).}
\item{normalized.by.row}{A matrix of normalized m/z values (by row, dividing by the maximum).}
}
\description{
This function processes Mass Spectrometry (MS) data, extracting mass and time values,
and generating several transformations of the MS data, including normalization.
}
\examples{
\dontrun{
  # Example usage:
  result <- transfer.ms(dat, mass.pattern = "m[0-9]", time.pattern = "time_MS")
}
}
