% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack.R
\name{Zeitkto}
\alias{Zeitkto}
\title{Convert Zeitkonto to Hours and Minutes}
\usage{
Zeitkto(Zeitkonto)
}
\arguments{
\item{Zeitkonto}{Numeric, the time account balance to convert.}
}
\value{
A string in the format "X h und Y Minuten".
}
\description{
This function converts a time account (Zeitkonto) in numeric format to hours and minutes in a readable format.
}
